Most of work done on acks. Now to wire it all together
