+++ clients/native/src/client/real_messages_control/acknowlegement_control/input_message_listener.rs
@@ -1,0 +1,121 @@
+// Copyright 2020 Nym Technologies SA
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::client::inbound_messages::{InputMessage, InputMessageReceiver};
+use crate::client::real_messages_control::acknowlegement_control::{
+    try_get_valid_topology_ref, PendingAcknowledgement, PendingAcksMap,
+};
+use crate::client::real_traffic_stream::RealSphinxSender;
+use crate::client::topology_control::TopologyAccessor;
+use futures::StreamExt;
+use log::*;
+use nymsphinx::acknowledgements::AckAes128Key;
+use nymsphinx::addressing::clients::Recipient;
+use nymsphinx::chunking::MessageChunker;
+use rand::{CryptoRng, Rng};
+use std::sync::Arc;
+use topology::NymTopology;
+
+// responsible for splitting received message and initial sending attempt
+// possible future TODO: the fields are IDENTICAL to the ones in RetransmissionRequestListener
+// perhaps some code could be shared?
+pub(super) struct InputMessageListener<R, T>
+where
+    R: CryptoRng + Rng,
+    T: NymTopology,
+{
+    ack_key: Arc<AckAes128Key>,
+    ack_recipient: Recipient,
+    input_receiver: InputMessageReceiver,
+    message_chunker: MessageChunker<R>,
+    pending_acks: PendingAcksMap,
+    real_sphinx_sender: RealSphinxSender,
+    topology_access: TopologyAccessor<T>,
+}
+
+impl<R, T> InputMessageListener<R, T>
+where
+    R: CryptoRng + Rng,
+    T: NymTopology,
+{
+    pub(super) fn new(
+        ack_key: Arc<AckAes128Key>,
+        ack_recipient: Recipient,
+        input_receiver: InputMessageReceiver,
+        message_chunker: MessageChunker<R>,
+        pending_acks: PendingAcksMap,
+        real_sphinx_sender: RealSphinxSender,
+        topology_access: TopologyAccessor<T>,
+    ) -> Self {
+        InputMessageListener {
+            ack_key,
+            ack_recipient,
+            input_receiver,
+            message_chunker,
+            pending_acks,
+            real_sphinx_sender,
+            topology_access,
+        }
+    }
+
+    async fn on_input_message(&mut self, msg: InputMessage) {
+        let (recipient, content) = msg.destruct();
+        let split_message = self.message_chunker.split_message(&content);
+        let topology_permit = &self.topology_access.get_read_permit().await;
+
+        let topology_ref_option =
+            try_get_valid_topology_ref(&self.ack_recipient, &recipient, topology_permit);
+        if topology_ref_option.is_none() {
+            warn!("Could not process the message - the network topology is invalid");
+            return;
+        }
+        let topology_ref = topology_ref_option.unwrap();
+
+        let mut pending_acks = Vec::with_capacity(split_message.len());
+
+        for message_chunk in split_message {
+            // since the paths can be constructed, this CAN'T fail, if it does, there's a bug somewhere
+            let frag_id = message_chunk.fragment_identifier();
+            // we need to clone it because we need to keep it in memory in case we had to retransmit
+            // it. And then we'd need to recreate entire ACK again.
+            let chunk_clone = message_chunk.clone();
+            let (total_delay, packet) = self
+                .message_chunker
+                .prepare_chunk_for_sending(chunk_clone, topology_ref, &self.ack_key, &recipient)
+                .unwrap();
+
+            self.real_sphinx_sender.unbounded_send(packet).unwrap();
+
+            let pending_ack =
+                PendingAcknowledgement::new(message_chunk, total_delay, recipient.clone());
+
+            pending_acks.push((frag_id, pending_ack));
+        }
+
+        let mut pending_acks_map_write_guard = self.pending_acks.write().await;
+        for (frag_id, pending_ack) in pending_acks.into_iter() {
+            if let Some(_) = pending_acks_map_write_guard.insert(frag_id, pending_ack) {
+                panic!("Tried to insert duplicate pending ack")
+            }
+        }
+    }
+
+    pub(super) async fn run(&mut self) {
+        debug!("Started InputMessageListener");
+        while let Some(input_msg) = self.input_receiver.next().await {
+            self.on_input_message(input_msg).await;
+        }
+        error!("TODO: error msg. Or maybe panic?")
+    }
+}
